Don't do a dynamic search for languages
Using Module::Pluggable to search dynamically for languages we
support was a really bad idea. Once Lingua::Boolean::Tiny was
released, we started loading those packages if they were
available, causing the mysterious "Not a HASH reference" errors
seen in CPANTesters.

To resolve this, I've hardcoded the two languages we support.

I will also be marking this module as deprecated. If you have
any notion that you want to continue using this module, you
should probably reconsider that, and if you still want to, then
use Lingua::Boolean::Tiny instead. This module was explictly an
experiment with the concept, and with API design, and is now
deprecated.